window-placement/multi-screen-window-open.tentative.https.html

<!DOCTYPE html>
<meta name="timeout" content="long">
<!-- user agents are not required to support open features other than `noopener`
     and on some platforms position and size features don't make sense -->
<meta name="flags" content="may">
<title>Window Management test: target-screen window.open()</title>
<link rel="help" href="https://w3c.github.io/window-management/">
Tests use of multi-screen details to open a popup window on each screen.<br>
The host device must have 2+ screens to yield meaningful results.<br><br>
<button id="closeButton" onclick="closeAll">Close any open popups</button><br>
<input id="autoClose" type="checkbox" checked=true>Auto-close popups</input>
<ul id="list"></ul>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/helpers.js"></script>
<script>
'use strict';

let popups = [];
function closeAll() {
  popups.forEach(p => p.close());
}

// Open a popup on screen `s`, assert expected placement, and return the popup.
async function openPopupOnScreen(s) {
  const left = s.availLeft + Math.floor(s.availWidth / 2) - 150;
  const top = s.availTop + Math.floor(s.availHeight / 2) - 50;
  const f = `left=${left},top=${top},width=300,height=100`;
  log(`Opening a popup with features '${f}' targeting ${screenLog(s)}`);
  // Window.open() synchronously returns a Window with estimated screenLeft|Top,
  // which may be clamped to the opener's screen or incompletely initialized.
  let p = window.open('/resources/blank.html', '', f);
  const initialBounds = windowLog(p);
  log(`<div style='margin-left: 40px'>Initial bounds: ${initialBounds}</div>`);

  // Assert the popup is eventually placed with the expected screen and bounds.
  // This may occur after window load, document ready and visible, etc.
  await poll(() => { return p.screenLeft == left && p.screenTop == top }); 
  p.document.write(`Requested: (${left},${top} 300x100) <br> \
       Initial: ${initialBounds} <br> \
       Resolved: ${windowLog(p)}`);
  log(`<div style='margin-left: 40px'>Resolved bounds: ${windowLog(p)}</div>`);
  const context = `popup: ${windowLog(p)}, ${screenLog(s)}`
  assert_equals(p.screenLeft, left, context);
  assert_equals(p.screenTop, top, context);
  return p;
}

promise_test(async setUpTest => {
  await setUpWindowManagement(setUpTest);
  for (const s of window.screenDetails.screens) {
    const name = `Open a popup on '${s.label}'`;
    await promise_test(async test => {
      await buttonClick(test, name);
      popups.push(await openPopupOnScreen(s));
      if (autoClose.checked)
        test.add_cleanup(closeAll);
    }, name);
  }
}, 'Use multi-screen details to open a popup window on each screen');
</script>
